(PC-14859)[PRO] refactor: offer v3: move EXTRA_DATA_FIELDS to core/Offers
